% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_grade_sum.R
\name{ae_plot_grade_sum}
\alias{ae_plot_grade_sum}
\alias{ae_plot_grade_n}
\title{Graphic representation of AEs}
\usage{
ae_plot_grade_sum(
  data_ae,
  ...,
  data_pat,
  low = "#ffc425",
  high = "#d11141",
  weights = NULL,
  arm = NULL,
  grade = "AEGR",
  subjid = "SUBJID"
)
}
\arguments{
\item{data_ae}{adverse event dataset, one row per AE, containing \code{subjid}, \code{grade}, \code{group1}, and potentially \code{group2}.}

\item{...}{unused}

\item{data_pat}{enrollment dataset, one row per patient, containing \code{subjid} (and \code{arm} if needed). All patients should be in this dataset.}

\item{low}{the color of Grade 1 AE}

\item{high}{the color of Grade 5 AE}

\item{weights}{(optional) a length 5 numeric vector, giving the weights of each grade}

\item{arm}{name of the treatment column in \code{data_pat}. Case-insensitive. Can be set to \code{NULL}.}

\item{grade}{name of the AE grade column in \code{data_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{data_ae} and \code{data_pat}. Case-insensitive.}
}
\value{
a ggplot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Produce an alternative graphic representation of AE, counting AE as bars for each patient, colored by grade. Can be faceted by treatment arm.
}
\examples{
tm = grstat_example()
attach(tm, warn.conflicts=FALSE)
ae_plot_grade_sum(data_ae=ae, data_pat=enrolres)
ae_plot_grade_sum(data_ae=ae, data_pat=enrolres, arm="ARM")
ae_plot_grade_sum(data_ae=ae, data_pat=enrolres, arm="ARM", weights=c(1,1,3,6,10))
}
\seealso{
\code{\link[=ae_table_grade]{ae_table_grade()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade]{ae_plot_grade()}}, \code{\link[=ae_plot_grade_sum]{ae_plot_grade_sum()}}, \code{\link[=butterfly_plot]{butterfly_plot()}}
}
