% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_table_grade_max}
\alias{ae_table_grade_max}
\title{Summary tables for AE by grade max}
\usage{
ae_table_grade_max(
  df_ae,
  ...,
  df_enrol,
  arm = "ARM",
  subjid = "SUBJID",
  soc = "AESOC",
  grade = "AEGR",
  total = TRUE,
  digits = 0
)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade.}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL}.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{soc}{name of the SOC column in \code{df_ae}. Case-insensitive. Grade will be considered 0 if missing (e.g. if patient if absent from \code{df_ae}).}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{total}{whether to add a \code{total} column for each arm.}

\item{digits}{significant digits for percentages.}
}
\value{
a crosstable (dataframe)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The function \code{ae_table_grade_max()} creates a summary table of the maximum AE grade experienced per each patient.
The resulting crosstable can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{

tm = edc_example_ae()
ae_table_grade_max(df_ae=tm$ae, df_enrol=tm$enrolres)
ae_table_grade_max(df_ae=tm$ae, df_enrol=tm$enrolres, arm=NULL)

\dontrun{
#you can use as_flextable() to get an HTML flextable
#you can use modificators modificators from the flextable package
library(flextable)
ae_table_grade_max(df_ae=tm$ae, df_enrol=tm$enrolres, arm=NULL) \%>\% 
  as_flextable() \%>\% 
  add_footer_lines("Percentages are given as the proportion of patients 
                    presenting at most one AE of given grade")
ae_table_grade_max(df_ae=tm$ae, df_enrol=tm$enrolres) \%>\%
  as_flextable(by_header="Both arms") \%>\% 
  highlight(i=~variable=="Grade 5", j=-1)
}
}
\seealso{
\code{\link[=ae_table_grade_max]{ae_table_grade_max()}}, \code{\link[=ae_table_grade_n]{ae_table_grade_n()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade_max]{ae_plot_grade_max()}}, \code{\link[=ae_plot_grade_n]{ae_plot_grade_n()}}
}
