% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_table_grade_n}
\alias{ae_table_grade_n}
\title{Summary tables for AE}
\usage{
ae_table_grade_n(
  df_ae,
  ...,
  df_enrol,
  arm = "ARM",
  grade = "AEGR",
  subjid = "SUBJID",
  soc = "AESOC",
  total = FALSE,
  digits = 0
)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade.}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL}.}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{soc}{name of the SOC column in \code{df_ae}. Case-insensitive. Grade will be considered 0 if missing (e.g. if patient if absent from \code{df_ae}).}

\item{total}{whether to add a \code{total} column for each arm.}

\item{digits}{significant digits for percentages.}
}
\value{
a crosstable
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
tm = grstat_example()

ae_table_grade_n(df_ae=tm$ae, df_enrol=tm$enrolres) \%>\%
  as_flextable() \%>\%
  flextable::add_footer_lines("Percentages are given as the proportion of patients
                               presenting at least one AE of given grade")

ae_table_grade_n(df_ae=tm$ae, df_enrol=tm$enrolres, arm=NULL) \%>\%
  as_flextable(by_header=F) \%>\%
  flextable::set_header_labels(values=c("","","N (\%)"))

#To get SAE only, filter df_ae first
tm$ae \%>\% filter(sae==TRUE) \%>\% ae_table_grade_n(df_enrol=tm$enrolres, arm=NULL)
}
}
\seealso{
\code{\link[=ae_table_grade_max]{ae_table_grade_max()}}, \code{\link[=ae_table_grade_n]{ae_table_grade_n()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade_max]{ae_plot_grade_max()}}, \code{\link[=ae_plot_grade_n]{ae_plot_grade_n()}}
}
