% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{as_flextable.ae_table_grade}
\alias{as_flextable.ae_table_grade}
\title{Convert AE Table to Flextable}
\usage{
\method{as_flextable}{ae_table_grade}(x, ..., padding_v = NULL)
}
\arguments{
\item{x}{An object of class \code{ae_table_grade}.}

\item{...}{Unused.}

\item{padding_v}{Vertical padding for cells.}
}
\value{
A \code{flextable} object ready to print or export.
}
\description{
Converts an object of class \code{ae_table_grade} to a formatted \code{flextable}.
}
