% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_template.R
\name{body_add_gr_sections_1_8}
\alias{body_add_gr_sections_1_8}
\title{Adds dummy sections to the report}
\usage{
body_add_gr_sections_1_8(doc, sections = 1:8)
}
\arguments{
\item{doc}{an officer \code{rdocx} object}

\item{sections}{Sections to add}
}
\value{
an officer \code{rdocx} object
}
\description{
Adds sections 1 to 8 to the report so that the first title points to the
standard report section 9 "Inclusion".
}
\examples{
officer::read_docx() \%>\%
  body_add_gr_sections_1_8() \%>\%
  crosstable::write_and_open()
}
