% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boin_plot.R
\name{boin_plot}
\alias{boin_plot}
\title{Plot BOIN decision boundaries and patient-level data}
\usage{
boin_plot(
  data_boin,
  data_patients = NULL,
  ...,
  doses,
  gantt_include = FALSE,
  gantt_labels = NULL
)
}
\arguments{
\item{data_boin}{A \code{boin} object from the \code{\link[BOIN:get.boundary]{BOIN::get.boundary()}} function \strong{or} a \code{data.frame}
with columns: \code{n_eval}, \code{escalate_if_inf}, \code{deescalate_if_sup}, \code{eliminate_if_sup}.}

\item{data_patients}{Optional data frame containing patient-level information,
with columns \code{subjid}, \code{dose} (character), and \code{dlt} (logical).
If \code{gantt_include==TRUE}, it must also include date of enrolment \code{date_enrol}
and date of followup end \code{date_end_fu}.}

\item{...}{Unused.}

\item{doses}{Named vector or list giving dose labels. Should match the order
of dose levels used in the BOIN design.}

\item{gantt_include}{Logical, whether to include a Gantt chart of follow-up.}

\item{gantt_labels}{Optional vector with labels as names and dates as values,
to display in the Gantt chart.}
}
\value{
A \code{ggplot2} object showing BOIN decision rules and patient data overlay (optional). If
\code{gantt_include==TRUE}, the output is a \code{patchwork} object with an additional Gantt diagram.
}
\description{
This function generates a plot of decision boundaries from a BOIN design,
optionally overlaid with patient-level data and a Gantt-style timeline.
}
\examples{
boin = BOIN::get.boundary(target=0.3, ncohort=6, cohortsize=1)
#In this example, 17 patients are included, 15 of which are already evaluated.
#Patients are deemed evaluable
set.seed(42)
data_patients = tibble(
  subjid = 1:17,
  dose = c("DL0", "DL0", "DL-1", "DL-1", "DL-1", "DL0", "DL0", "DL0",
           "DL1", "DL1", "DL1", "DL1", "DL1", "DL1", "DL1", "DL1", "DL1"),
  dlt = c(FALSE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,
          FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, NA, NA),
  date_enrol = as.Date("2025-01-01") +(1:17)*30 + runif(17, -13, 13),
  date_end_fu = date_enrol+30
)

#default
boin_plot(boin,
          doses=c("DL-1", "DL0", "DL1", "DL2"))
#with patient labels
boin_plot(boin, data_patients=data_patients,
          doses=c("DL-1", "DL0", "DL1", "DL2"))
#with gantt diagram
p = boin_plot(boin, data_patients=data_patients, gantt_include=TRUE,
              doses=c("DL-1", "DL0", "DL1", "DL2"))
p
#with post production
p[[1]] = p[[1]] + scale_color_manual(values=c("red", "green", "blue", "yellow"))
p[[2]] = p[[2]] + scale_color_manual(values=c("red", "green", "blue"))
p

#with gantt labels
gantt_labels=c("Plot cutoff date"=as.Date("2026-08-25"),
               "Dose reevaluation"=as.Date("2025-08-25"))

boin_plot(data_boin=boin, data_patients=data_patients,
          doses = c("DL-1", "DL0", "DL1", "DL2"),
          gantt_include=TRUE,
          gantt_labels=gantt_labels)
}
