% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_soc.R
\name{butterfly_plot}
\alias{butterfly_plot}
\alias{ae_plot_soc}
\title{Graphic representation of AEs by CTCAE SOC}
\usage{
butterfly_plot(
  df_ae,
  ...,
  df_enrol,
  severe = NULL,
  sort_by = c("total", "severe"),
  range_min = NULL,
  arm = "ARM",
  subjid = "SUBJID",
  soc = "AESOC"
)

ae_plot_soc(df_ae, ..., df_enrol, severe, sort_by, range_min, arm, subjid, soc)
}
\arguments{
\item{...}{unused}

\item{severe}{name of the logical column in \code{df_ae} telling whether an AE is severe. Case-insensitive.}

\item{sort_by}{either "total" or "severe"}

\item{range_min}{The minimum value for the upper limit of the x-axis range. Set to \code{1} to always include 100\%.}

\item{arm}{name of the treatment column in \code{data_pat}. Case-insensitive. Can be set to \code{NULL}.}

\item{subjid}{name of the patient ID in both \code{data_ae} and \code{data_pat}. Case-insensitive.}
}
\value{
a crosstable (dataframe)
}
\description{
Produces a graphic representation of AEs by CTCAE SOC.
}
\examples{

tm = grstat_example(N=100)
attach(tm, warn.conflicts=FALSE)

ae2 = ae \%>\%
  dplyr::mutate(serious = sae=="Yes")

butterfly_plot(ae2, df_enrol=enrolres, range_min=0.5)
butterfly_plot(ae2, df_enrol=head(enrolres,9), range_min=0.5)

ae2 \%>\%
  butterfly_plot(df_enrol=enrolres, severe="serious") +
  ggplot2::labs(caption="Darker areas represent Serious Adverse Events")
}
\seealso{
\code{\link[=ae_table_grade]{ae_table_grade()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade]{ae_plot_grade()}}, \code{\link[=ae_plot_grade_sum]{ae_plot_grade_sum()}}, \code{\link[=butterfly_plot]{butterfly_plot()}}
}
