% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{check_constancy}
\alias{check_constancy}
\title{Target Lesion site and evaluation method should have one value per patient
Responses should have one value per date}
\usage{
check_constancy(rc)
}
\description{
Target Lesion site and evaluation method should have one value per patient
Responses should have one value per date
}
\keyword{internal}
