% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{check_recist}
\alias{check_recist}
\title{Check RECIST data}
\usage{
check_recist(rc, mapping = gr_recist_mapping())
}
\arguments{
\item{rc}{the recist dataset}

\item{mapping}{the result of \code{\link[=gr_recist_mapping]{gr_recist_mapping()}}}
}
\value{
a tibble of nested checks
}
\description{
Check RECIST datasets
}
\seealso{
\href{https://ctep.cancer.gov/protocoldevelopment/docs/recist_guideline.pdf}{RECIST guidelines}
}
