% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{check_target_response}
\alias{check_target_response}
\title{Check impossible cases for Target Lesions response}
\usage{
check_target_response(rc, rc_short)
}
\description{
Check impossible cases for Target Lesions response
}
\keyword{internal}
