% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisation_modeles.R
\name{cox_regtable}
\alias{cox_regtable}
\title{Results table from a Cox Model}
\usage{
cox_regtable(fit)
}
\arguments{
\item{fit}{an object of class coxph, output of the survival::coxph function}
}
\value{
an object of class regtable
}
\description{
Results table from a Cox Model
}
