% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_fu}
\alias{example_fu}
\title{Generate Simulated Survival Data}
\usage{
example_fu(
  enrolres,
  recist,
  seed,
  fu_lambda_censor = 0.2,
  fu_lambda_control = 0.2,
  fu_beta_arm = -0.6,
  fu_beta_prog_status = 2,
  ...
)
}
\arguments{
\item{enrolres}{the enrolment result table, from \code{\link[=example_enrol]{example_enrol()}}.}

\item{recist}{the recist table with progression date, from \code{\link[=example_rc]{example_rc()}}.}

\item{seed}{Integer. Random seed for reproducibility (can be \code{NULL}).}

\item{fu_lambda_censor}{numeric. Rate parameter of an exponential distribution used to simulate censoring times.}

\item{fu_lambda_control}{numeric. Scale parameter of an exponential baseline hazard function.}

\item{fu_beta_arm}{numeric. Parameter associated to the treatment effect.}

\item{fu_beta_prog_status}{numeric. Parameter associated to a known progression on a RECIST assessment.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A tibble with \code{N} rows and the following columns:
\itemize{
\item \code{subjid}: Subject ID
\item \code{pfs_status}: Progression-free survival status of the patient
\item \code{pfs_date}: Date of latest news before progression/death
\item \code{os_status}: Overall survival status of the patient (0 - Alive/Censored , 1 - Dead)
\item \code{os_date}: Date of latest news (death date if dead, otherwise censoring date)
}
}
\description{
Internal function that simulates survival data, which depend on treatment arm
and date of progression if applicable.
}
\keyword{internal}
