% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_rc}
\alias{example_rc}
\title{Simulate a RECIST dataset}
\usage{
example_rc(
  enrolres,
  seed,
  rc_num_timepoints = 5,
  rc_p_new_lesions = 0.09,
  rc_p_not_evaluable = 0.01,
  rc_p_nt_lesions_yn = 0.5,
  rc_p_nt_lesions_resp = list(CR = 0.72, SD = 0.24, PD = 0.01, NE = 0.01),
  rc_sd_tlsum_noise = 0.5,
  rc_coef_treatement = 3,
  ...
)
}
\arguments{
\item{enrolres}{the enrolment result table, from \code{\link[=example_enrol]{example_enrol()}}.}

\item{seed}{Integer. Random seed for reproducibility (can be \code{NULL}).}

\item{rc_num_timepoints}{Integer. Number of timepoints for each patient, including baseline.}

\item{rc_p_new_lesions}{Integer. Probability of a new lesion}

\item{rc_p_not_evaluable}{Integer. Probability of a Not Evaluable measure}

\item{rc_p_nt_lesions_yn}{Integer. Probability of having Non-Target Lesions}

\item{rc_p_nt_lesions_resp}{Integer list. Probability of each Non-Target Lesions response, if present}

\item{rc_sd_tlsum_noise}{Integer. Standard deviation for the evolution of the Target Lesion sum of width}

\item{rc_coef_treatement}{Integer. Differentiates the difference in effect between the control and treatment arm (2 arms only). For example, \code{rc_coef_treatement} = 2 mean that the growth rate of the tumor is divide per 2 and the elimination rate is multiplied per 2. Also, the probability of a new lesion is multiplied by 2.}
}
\value{
A tibble with \code{N} rows and the following columns:
\itemize{
\item \code{subjid}: The patient identifier
\item \code{rcvisit}: The visit number
\item \code{rcdt}: The visit date
\item \code{rctlsum}: The Target Lesion length sum at each time point. The evolution of the value is calculated based on the percentage variation in tumor size from the previous time point. This variation is simulated using a uniform distribution between -30 and 30, with added noise (the noise follows a normal distribution with a mean of 0 and a standard deviation of \code{rc_sd_tlsum_noise}).
\item \code{rctlresp}: The response associated with Target Lesions
\item \code{rcntlresp}: The response associated with Non-Target Lesions
\item \code{rcnew}: The appearance of a new lesion
\item \code{rcresp}: The global RECIST response
}
}
\description{
Internal function that simulates a synthetic RECIST dataset following the
conventions of clinical oncology trials. In the simulation, Target Lesions
response depends on the treatment arm \code{enrolres$arm} (not \code{enrolres$arm3}).
}
\keyword{internal}
