% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{gr_recist_mapping}
\alias{gr_recist_mapping}
\title{Default mapping for RECIST dataset}
\usage{
gr_recist_mapping()
}
\value{
a named character vector
}
\description{
Generates the default mapping for the RECIST dataset. Values from \code{subjid}
to \code{global_resp} are mandatory. Other values will only trigger specified
checks if set. For \code{target_is_node}, if not set, it will be guessed from
the presence of "node" in \code{target_site}.
}
\examples{
gr_recist_mapping()
}
