% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{grstat_example}
\alias{grstat_example}
\title{Example databases}
\usage{
grstat_example(N = 50, seed = 42)
}
\arguments{
\item{N}{the number of patients}

\item{seed}{the random seed}
}
\description{
Example tables, mostly used in examples and tests.
}
