% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{rc_check_derived_columns}
\alias{rc_check_derived_columns}
\title{Checks on values calculated in the EDC software
Check target sum at any time, at baseline, and at Nadir
In TrialMaster, if \code{target_sum} is modified, \code{target_sum_bl} & \code{target_sum_min}
are not updated automatically.
This also makes duplicates possible.}
\usage{
rc_check_derived_columns(rc)
}
\description{
Checks on values calculated in the EDC software
Check target sum at any time, at baseline, and at Nadir
In TrialMaster, if \code{target_sum} is modified, \code{target_sum_bl} & \code{target_sum_min}
are not updated automatically.
This also makes duplicates possible.
}
\keyword{internal}
