% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{rc_check_global_response}
\alias{rc_check_global_response}
\title{Check impossible cases for Target Lesions response}
\usage{
rc_check_global_response(rc_short)
}
\description{
Check impossible cases for Target Lesions response
}
\keyword{internal}
