% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{rc_check_nontarget_response}
\alias{rc_check_nontarget_response}
\title{Check impossible cases for Non-Target Lesions response}
\usage{
rc_check_nontarget_response(rc)
}
\description{
Check impossible cases for Non-Target Lesions response
}
\keyword{internal}
