% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{recist_report}
\alias{recist_report}
\title{Create a RECIST check HTML report}
\usage{
recist_report(recist_check, output_file = "recist_check.html", open = TRUE)
}
\arguments{
\item{recist_check}{the output of \code{\link[=check_recist]{check_recist()}}}

\item{output_file}{the HTML report file name.}

\item{open}{whether to open the report afterward.}
}
\value{
\code{output_file} invisibly. Called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Turn the table created by \code{\link[=check_recist]{check_recist()}} into an HTML report.
}
\examples{
⁠\dontrun{
 #we unfortunately cannot provide a flawed simulated recist dataset
 mapping = gr_recist_mapping()
 recist_check = check_recist(rc, mapping=mapping)
 recist_check
 recist_report(recist_check)
}⁠

}
