% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{recist_report_xlsx}
\alias{recist_report_xlsx}
\title{Create a RECIST check HTML report}
\usage{
recist_report_xlsx(
  recist_check,
  output_file = "recist_check.xlsx",
  open = TRUE
)
}
\arguments{
\item{recist_check}{the output of \code{\link[=check_recist]{check_recist()}}}

\item{output_file}{the HTML report file name.}

\item{open}{whether to open the report afterward.}

\item{title}{the HTML report title.}
}
\value{
\code{output_file} invisibly. Called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Turn the table created by \code{\link[=check_recist]{check_recist()}} into an HTML report.
}
\examples{
# we unfortunately cannot provide a flawed simulated recist dataset, at least not yet
\dontrun{
db = read_database()
mapping = gr_recist_mapping()
recist_check = check_recist(db$rc, mapping=mapping)
recist_check
recist_report_html(recist_check, title="RECIST Check - Study XXX")
recist_report_xlsx(recist_check)
}

}
