% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardisation_modeles.R
\name{as_flextable.regtable}
\alias{as_flextable.regtable}
\title{Turns a \code{regtable} object into a formatted \code{flextable}}
\usage{
\method{as_flextable}{regtable}(x, ..., padding_v = 10)
}
\arguments{
\item{x}{a dataframe, resulting of \code{cox_regtable()}}

\item{...}{unused}

\item{padding_v}{a numeric of lenght up to 2, giving the vertical padding of body (1) and header (2)}
}
\value{
a formatted flextable
}
\description{
Turns a \code{regtable} object into a formatted \code{flextable}
}
