% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimmerplot.R
\name{save_plotly}
\alias{save_plotly}
\title{Save a plotly to an HTML file}
\usage{
save_plotly(p, file, ...)
}
\arguments{
\item{p}{a plot object (\code{plotly} or \code{ggplot})}

\item{file}{a file path to save the HTML file}

\item{...}{passed on to \link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget}}
}
\description{
Save a plotly to an HTML file
}
\examples{
\dontrun{
tm = edc_example_plot()
p = edc_swimmerplot(tm$.lookup, id_lim=c(5,45))
save_plotly(p, "graph/swimplots/edc_swimmerplot.html", title="My Swimmerplot")
}
}
