% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit_stack.R
\name{survfit_stack}
\alias{survfit_stack}
\title{Stack multiple survival endpoints into a single \code{survfit2} object}
\usage{
survfit_stack(data, surv_list)
}
\arguments{
\item{data}{A data frame containing the survival variables.}

\item{surv_list}{A named list where each element is a character vector of:
\itemize{
\item Length 2: \code{time}, \code{event} (for standard right-censored survival)
\item Length 3: \code{time1}, \code{time2}, \code{event} (for interval-censored survival and left truncation)
}

The names of the list elements will be used as a grouping variable in the output.}
}
\value{
A \code{survfit} object (or \code{survfit2} if \code{ggsurvfit} is installed)
}
\description{
This function combines several survival endpoints from the same dataset by stacking them
into a long format and fitting a single \code{survfit} object.
It does \strong{not} support competing risks models (i.e., when the event variable is a factor).
}
\examples{
library(dplyr)
set.seed(42)
df_surv = survival::lung \%>\%
  as_tibble() \%>\%
  mutate(
    time_os = time,
    event_os = status == 2,
    time_pfs = time / (1 + runif(n())),
    event_pfs = ifelse(runif(n()) > 0.3, event_os, 1),
    time_efs = time_pfs / (1 + runif(n())),
    event_efs = ifelse(runif(n()) > 0.3, event_pfs, 1)
  )

surv_list = list(
  "Overall survival" = c("time_os", "event_os"),
  "Progression-free survival" = c("time_pfs", "event_pfs"),
  "Event-free survival" = c("time_efs", "event_efs")
)

s = survfit_stack(df_surv, surv_list=surv_list)
s

if(require("ggsurvfit")){
  s \%>\%
    ggsurvfit() +
    add_confidence_interval() +
    add_risktable() +
    scale_ggsurvfit()

}
}
