% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_table_grade}
\alias{ae_table_grade}
\title{Summary tables for AE}
\usage{
ae_table_grade(
  df_ae,
  ...,
  df_enrol,
  variant = c("max", "sup", "eq"),
  arm = NULL,
  grade = "AEGR",
  subjid = "SUBJID",
  percent = TRUE,
  total = TRUE
)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade.}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{variant}{one or several of \code{c("max", "sup", "eq")}. \code{max} computes the maximum AE grade per patient, \code{sup} computes the number of patients having experienced at least one AE of grade higher or equal to X, and \code{eq} computes the number of patients having experienced at least one AE of grade equal to X.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL}.}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{percent}{whether to show percentages with counts. Defaults to TRUE. Can also be "only" to not show counts.}

\item{total}{whether to add a \code{total} column for each arm.}
}
\value{
a crosstable
}
\description{
Summary tables for AE
}
\examples{
tm = edc_example_ae()
load_list(tm)

if(require(flextable)){

ae_table_grade(df_ae=ae, df_enrol=enrolres, arm=NULL) \%>\% 
  as_flextable(header_show_n=TRUE)

ae_table_grade(df_ae=ae, df_enrol=enrolres, arm="ARM") \%>\% 
  as_flextable(header_show_n=TRUE) 
  
#To get SAE only, filter df_ae first
library(dplyr)
ae \%>\% 
  filter(sae=="Yes") \%>\% 
  ae_table_grade(df_enrol=enrolres, arm="ARM") \%>\% 
  dplyr::mutate_all(~stringr::str_replace(.x, "AE", "SAE")) \%>\% 
  as_flextable(header_show_n=TRUE) 
  
}
}
\seealso{
\code{\link[=ae_table_grade_max]{ae_table_grade_max()}}, \code{\link[=ae_table_grade_n]{ae_table_grade_n()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade_max]{ae_plot_grade_max()}}, \code{\link[=ae_plot_grade_n]{ae_plot_grade_n()}}
}
