% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_grade.R
\name{ae_table_grade}
\alias{ae_table_grade}
\title{Summary tables for AE}
\usage{
ae_table_grade(
  df_ae,
  ...,
  df_enrol,
  variant = c("max", "sup", "eq"),
  arm = NULL,
  grade = "AEGR",
  subjid = "SUBJID",
  ae_label = "AE",
  percent = TRUE,
  digits = 2,
  total = TRUE
)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade.}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{variant}{one or several of \code{c("max", "sup", "eq")}. \code{max} computes the maximum AE grade per patient, \code{sup} computes the number of patients having experienced at least one AE of grade higher or equal to X, and \code{eq} computes the number of patients having experienced at least one AE of grade equal to X.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL}.}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{ae_label}{the label of adverse events, usually "AE" or "SAE".}

\item{percent}{whether to show percentages with counts. Defaults to TRUE. Can also be "only" to not show counts.}

\item{digits}{significant digits for percentages.}

\item{total}{whether to add a \code{total} column for each arm.}
}
\value{
a crosstable
}
\description{
Summary tables for AE
}
\examples{
tm = grstat_example()
attach(tm, warn.conflicts=FALSE)

ae_table_grade(df_ae=ae, df_enrol=enrolres, arm=NULL) \%>\%
  as_flextable(header_show_n=TRUE)

ae_table_grade(df_ae=ae, df_enrol=enrolres, arm="arm") \%>\%
  as_flextable(header_show_n=TRUE)

#To get SAE only, filter df_ae first
ae \%>\%
  dplyr::filter(sae=="Yes") \%>\%
  ae_table_grade(df_enrol=enrolres, arm="arm", ae_label="SAE") \%>\%
  as_flextable(header_show_n=TRUE)

#To describe a sub-population, filter df_enrol first
enrolres2 = enrolres \%>\%
  dplyr::filter(arm=="Control")
ae \%>\%
  ae_table_grade(df_enrol=enrolres2, arm="arm") \%>\%
  as_flextable(header_show_n=TRUE)

#You can also filter the AE table
ae \%>\%
  ae_table_grade(df_enrol=enrolres, arm="arm") \%>\%
  dplyr::filter(!variable \%in\% c("Grade 1", "Grade 2")) \%>\%
  as_flextable(header_show_n=TRUE)
}
\seealso{
\code{\link[=ae_table_grade]{ae_table_grade()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade]{ae_plot_grade()}}, \code{\link[=ae_plot_grade_sum]{ae_plot_grade_sum()}}, \code{\link[=butterfly_plot]{butterfly_plot()}}
}
