% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{ae_table_soc}
\alias{ae_table_soc}
\alias{as_flextable.ae_table_soc}
\title{Summary tables for AE by SOC}
\usage{
ae_table_soc(
  data_ae,
  ...,
  data_pat,
  measure = c("max", "sup", "eq"),
  group1 = "AESOC",
  group2 = NULL,
  arm = NULL,
  cols = c(grade = "AEGR", subjid = "SUBJID"),
  ae_groups = NULL,
  ae_label = "AE",
  sort_by_count = TRUE,
  total = TRUE,
  showNA = TRUE,
  digits = 0,
  warn_miss = FALSE
)

\method{as_flextable}{ae_table_soc}(
  x,
  ...,
  show_footer = c("both", "explanation", "example", "none"),
  arm_colors = c("#f2dcdb", "#dbe5f1", "#ebf1dd", "#e5e0ec"),
  padding_v = NULL
)
}
\arguments{
\item{data_ae}{adverse event dataset, one row per AE, containing \code{subjid}, \code{grade}, \code{group1}, and potentially \code{group2}.}

\item{...}{unused}

\item{data_pat}{enrollment dataset, one row per patient, containing \code{subjid} (and \code{arm} if needed). All patients should be in this dataset.}

\item{measure}{one or several of \code{c("max", "sup", "eq")}. \code{max} computes the maximum AE grade per patient, \code{sup} computes the number of patients having experienced at least one AE of grade higher or equal to X, and \code{eq} computes the number of patients having experienced at least one AE of grade equal to X.}

\item{group1, group2}{name of the 1st and 2nd order grouping columns in \code{data_ae}. Case-insensitive. Use labels for the flextable output. Usually, \code{group1} is the SOC and \code{group2} the term, but it can be any other grouping variable. \code{group2} can be set to \code{NULL} if not needed.}

\item{arm}{name of the treatment column in \code{data_pat}. Case-insensitive. Can be set to \code{NULL}.}

\item{cols}{a named character vector mapping column names. Should contain at least \code{grade} and \code{subjid}. Case-insensitive.}

\item{ae_groups}{a named list specifying the grade values for each group.}

\item{ae_label}{Label used in the output tables (e.g. "AE", "SAE", "Toxicity").}

\item{sort_by_count}{whether to sort by the number of AE or by \code{group1} alphabetically.}

\item{total}{whether to add a \code{total} column for each arm.}

\item{showNA}{whether to display missing grades. Only relevant if \code{ae_groups} is not used.}

\item{digits}{significant digits for percentages.}

\item{warn_miss}{whether to warn for missing values.}

\item{x}{a dataframe, resulting of \code{ae_table_soc()}}

\item{arm_colors}{colors for the arm groups}

\item{padding_v}{a numeric of lenght up to 2, giving the vertical padding of body (1) and header (2)}
}
\value{
a dataframe (\code{ae_table_soc()}) or a flextable (\code{as_flextable()}).

a formatted flextable
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}\cr
The function \code{ae_table_soc()} creates a summary table of AE grades for each patient by group (usually according CTCAE SOC or term).
The resulting dataframe can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{
tm = grstat_example()
attach(tm, warn.conflicts=FALSE)

#Default
ae_table_soc(data_ae=ae, data_pat=enrolres) \%>\%
  as_flextable()

#By arm, with alternative measure
ae_table_soc(data_ae=ae, data_pat=enrolres, arm="arm", measure="sup", total=FALSE) \%>\%
  as_flextable()

#Sub-population, without footer
ae_table_soc(data_ae=ae, data_pat=head(enrolres, 10), arm="arm") \%>\%
  as_flextable(show_footer="none")

#Grouping grades, with only TERM
ae_groups = list("Any grade"=c(1:5,NA), "Grade 1-2"=1:2, "Grade 3-5"=3:5)
ae_table_soc(data_ae=ae, data_pat=head(enrolres, 10), group1="AETERM", ae_groups=ae_groups) \%>\%
  as_flextable()

#Stratified by both SOC and TERM
ae_table_soc(data_ae=ae, data_pat=head(enrolres, 10), arm="arm", group1="AESOC", group2="AETERM") \%>\%
  filter(!is.na(group2)) \%>\% #remove missing term
  as_flextable()
}
\seealso{
\code{\link[=ae_table_grade]{ae_table_grade()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade]{ae_plot_grade()}}, \code{\link[=ae_plot_grade_sum]{ae_plot_grade_sum()}}, \code{\link[=butterfly_plot]{butterfly_plot()}}
}
