% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{ae_table_soc}
\alias{ae_table_soc}
\alias{as_flextable.ae_table_soc}
\title{Summary tables for AE by SOC}
\usage{
ae_table_soc(
  df_ae,
  ...,
  df_enrol,
  variant = c("max", "sup", "eq"),
  arm = NULL,
  term = NULL,
  sort_by_count = TRUE,
  total = TRUE,
  showNA = TRUE,
  digits = 0,
  warn_miss = FALSE,
  grade = "AEGR",
  soc = "AESOC",
  subjid = "SUBJID"
)

\method{as_flextable}{ae_table_soc}(
  x,
  ...,
  arm_colors = c("#f2dcdb", "#dbe5f1", "#ebf1dd", "#e5e0ec"),
  padding_v = NULL
)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade.}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{variant}{one or several of \code{c("max", "sup", "eq")}. \code{max} computes the maximum AE grade per patient, \code{sup} computes the number of patients having experienced at least one AE of grade higher or equal to X, and \code{eq} computes the number of patients having experienced at least one AE of grade equal to X.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL}.}

\item{term}{name of the the CTCAE term column in \code{df_ae}. Case-insensitive. Can be set to \code{NULL}.}

\item{sort_by_count}{should the table be sorted by the number of AE or by SOC alphabetically.}

\item{total}{whether to add a \code{total} column for each arm.}

\item{showNA}{whether to display missing grades.}

\item{digits}{significant digits for percentages.}

\item{warn_miss}{whether to warn for missing values.}

\item{grade}{name of the AE grade column in \code{df_ae}. Case-insensitive.}

\item{soc}{name of the SOC column in \code{df_ae}. Case-insensitive. Grade will be considered 0 if missing (e.g. if patient if absent from \code{df_ae}).}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{x}{a dataframe, resulting of \code{ae_table_soc()}}

\item{arm_colors}{colors for the arm groups}

\item{padding_v}{a numeric of lenght up to 2, giving the vertical padding of body (1) and header (2)}
}
\value{
a dataframe (\code{ae_table_soc()}) or a flextable (\code{as_flextable()}).

a formatted flextable
}
\description{
The function \code{ae_table_soc()} creates a summary table of maximum AE grades for each patient according to term and SOC CTCAE.
The resulting dataframe can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{
tm = grstat_example()
attach(tm, warn.conflicts=FALSE)

ae_table_soc(df_ae=ae, df_enrol=enrolres)
ae_table_soc(df_ae=ae, df_enrol=enrolres, arm="arm")

#sub population
ae_table_soc(df_ae=ae, df_enrol=head(enrolres, 10), arm="arm")

#the resulting flextable can be customized using the flextable package
library(flextable)
ae_table_soc(ae, df_enrol=enrolres, total=FALSE) \%>\%
  as_flextable() \%>\%
  hline(i=~soc=="" & soc!=dplyr::lead(soc))
ae_table_soc(ae, df_enrol=enrolres, term=NULL, sort_by_count=FALSE) \%>\%
  as_flextable() \%>\%
  bold(i=~soc=="Eye disorders")
ae_table_soc(ae, df_enrol=enrolres, term=NULL, arm=NULL) \%>\%
  as_flextable() \%>\%
  highlight(i=~soc=="Hepatobiliary disorders", j="all_patients_Tot")
}
\seealso{
\code{\link[=ae_table_grade]{ae_table_grade()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade]{ae_plot_grade()}}, \code{\link[=ae_plot_grade_sum]{ae_plot_grade_sum()}}, \code{\link[=butterfly_plot]{butterfly_plot()}}
}
