% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_recist_rates.R
\name{aggregate_recist_rates}
\alias{aggregate_recist_rates}
\alias{as_flextable.aggregate_recist_rates}
\title{Calculate BOR}
\usage{
aggregate_recist_rates(data, ..., derived_endpoints = c("ORR", "CBR", "DCR"))

\method{as_flextable}{aggregate_recist_rates}(x, ...)
}
\arguments{
\item{data}{A dataset containing longitudinal RECIST data in long format.}

\item{...}{unused}

\item{derived_endpoints}{Character; Derived endpoints to compute from BOR. One or several of c("ORR", "CBR", "DCR"). See vignette("BOR") for endpoint definitions.}

\item{x}{a dataframe, resulting of \code{aggregate_recist_rates()}}
}
\value{
a dataframe (\code{aggregate_recist_rates()}) or a flextable (\code{as_flextable()}).

a formatted flextable
}
\description{
The function \code{aggregate_recist_rates()} creates a summary table of recist for each possible response.
The resulting dataframe can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{
recist = grstat_example()$recist
recist \%>\%
 calc_best_response(rc_resp = "rcresp", rc_date = "rcdt",
                    subjid = "subjid", rc_sum = "rctlsum", confirmed = FALSE) \%>\%
 aggregate_recist_rates(derived_endpoints=c("ORR", "CBR", "DCR")) \%>\%
 as_flextable()
#It is also possible to use the confirmation method for the ORR
recist \%>\%
 calc_best_response(rc_resp = "rcresp", rc_date = "rcdt",
                    subjid = "subjid", rc_sum = "rctlsum", confirmed = TRUE) \%>\%
 aggregate_recist_rates(derived_endpoints=c("ORR") \%>\%
 as_flextable()

}
