% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_soc.R
\name{butterfly_plot}
\alias{butterfly_plot}
\alias{ae_plot_soc}
\title{Graphic representation of AEs by CTCAE SOC}
\usage{
butterfly_plot(
  data_ae,
  ...,
  data_pat,
  severe = NULL,
  sort_by = c("total", "severe"),
  range_min = NULL,
  arm = "ARM",
  subjid = "SUBJID",
  group = "AESOC"
)

ae_plot_soc(data_ae, ..., data_pat, severe, sort_by, range_min, arm, subjid, group)
}
\arguments{
\item{data_ae}{adverse event dataset, one row per AE, containing \code{subjid}, \code{grade}, \code{group1}, and potentially \code{group2}.}

\item{...}{unused}

\item{data_pat}{enrollment dataset, one row per patient, containing \code{subjid} (and \code{arm} if needed). All patients should be in this dataset.}

\item{severe}{name of the logical column in \code{data_ae} telling whether an AE is severe. Case-insensitive.}

\item{sort_by}{either "total" or "severe"}

\item{range_min}{The minimum value for the upper limit of the x-axis range. Set to \code{1} to always include 100\%.}

\item{arm}{name of the treatment column in \code{data_pat}. Case-insensitive. Can be set to \code{NULL}.}

\item{subjid}{name of the patient ID in both \code{data_ae} and \code{data_pat}. Case-insensitive.}

\item{group}{name of the grouping column in \code{data_ae}. Case-insensitive. Usually, the SOC. This is the variable that will be plotted on the y-axis.}
}
\value{
a crosstable (dataframe)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}\cr
Produces a graphic representation of AEs by CTCAE SOC.
}
\examples{

tm = grstat_example(N=100)
attach(tm, warn.conflicts=FALSE)

ae2 = ae \%>\%
  dplyr::mutate(serious = sae=="Yes")

butterfly_plot(ae2, data_pat=enrolres, range_min=0.5)
butterfly_plot(ae2, data_pat=head(enrolres,9), range_min=0.5)

ae2 \%>\%
  butterfly_plot(data_pat=enrolres, severe="serious") +
  ggplot2::labs(caption="Darker areas represent Serious Adverse Events")
}
\seealso{
\code{\link[=ae_table_grade]{ae_table_grade()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade]{ae_plot_grade()}}, \code{\link[=ae_plot_grade_sum]{ae_plot_grade_sum()}}, \code{\link[=butterfly_plot]{butterfly_plot()}}
}
