% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table_soc.R
\name{butterfly_plot}
\alias{butterfly_plot}
\alias{ae_plot_soc}
\title{Graphic representation of AEs by soc (Butterfly plot)}
\usage{
butterfly_plot(
  df_ae,
  ...,
  df_enrol,
  severe = NULL,
  sort_by = c("total", "severe"),
  range_min = NULL,
  arm = "ARM",
  subjid = "SUBJID",
  soc = "AESOC"
)

ae_plot_soc(df_ae, ..., df_enrol, severe, sort_by, range_min, arm, subjid, soc)
}
\arguments{
\item{df_ae}{adverse event dataset, one row per AE, containing subjid, soc, and grade.}

\item{...}{unused}

\item{df_enrol}{enrollment dataset, one row per patient, containing subjid (and arm if needed). All patients should be in this dataset.}

\item{severe}{name of the logical column in \code{df_ae} telling whether an AE is severe. Case-insensitive.}

\item{sort_by}{either "total" or "severe"}

\item{range_min}{The minimum value for the upper limit of the x-axis range. Set to \code{1} to always include 100\%.}

\item{arm}{name of the treatment column in \code{df_enrol}. Case-insensitive. Can be set to \code{NULL}.}

\item{subjid}{name of the patient ID in both \code{df_ae} and \code{df_enrol}. Case-insensitive.}

\item{soc}{name of the SOC column in \code{df_ae}. Case-insensitive. Grade will be considered 0 if missing (e.g. if patient if absent from \code{df_ae}).}
}
\value{
a crosstable (dataframe)
}
\description{
Produces a graphic representation of AE, counting the maximum grade each patient experienced, colored by treatment arm. Returns up to 3 representations if \code{arm!=NULL}.
}
\details{
The function \code{butterfly_plot()} creates a summary table of the maximum AE grade experienced per each patient.
The resulting crosstable can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{

tm = edc_example_ae(N=100)
load_list(tm)

ae2 = ae \%>\% 
  dplyr::mutate(serious = sae=="Yes")
  
ae2 \%>\% 
  butterfly_plot(df_enrol=enrolres, range_min=0.5)
  
ae2 \%>\% 
  butterfly_plot(df_enrol=enrolres, severe="serious") + 
  ggplot2::labs(caption="Darker areas represent Serious Adverse Events")
}
\seealso{
\code{\link[=ae_table_grade_max]{ae_table_grade_max()}}, \code{\link[=ae_table_grade_n]{ae_table_grade_n()}}, \code{\link[=ae_table_soc]{ae_table_soc()}}, \code{\link[=ae_plot_grade_max]{ae_plot_grade_max()}}, \code{\link[=ae_plot_grade_n]{ae_plot_grade_n()}}
}
