% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{check_recist}
\alias{check_recist}
\title{Check a RECIST dataset}
\usage{
check_recist(rc, mapping = gr_recist_mapping(), exclude_post_pd = TRUE)
}
\arguments{
\item{rc}{The recist dataset to check}

\item{mapping}{The character vector defining the variable mapping. Refer to \code{\link[=gr_recist_mapping]{gr_recist_mapping()}} for default values and adjust as needed.}

\item{exclude_post_pd}{Logical; if \code{TRUE} (default), assessments after the first PD are excluded.}
}
\value{
a tibble of nested checks, of class \code{check_recist}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Perform multiple checks on a RECIST dataset. The checks can be exported as an HTML or Excel report using \code{recist_report_html()} or \code{recist_report_xlsx()}.
}
\examples{
# we unfortunately cannot provide a flawed simulated recist dataset, at least not yet
\dontrun{
db = read_database()
mapping = gr_recist_mapping()
recist_check = check_recist(db$rc, mapping=mapping)
recist_check
recist_report_html(recist_check, title="RECIST Check - Study XXX")
recist_report_xlsx(recist_check)
}

}
\seealso{
\href{https://ctep.cancer.gov/protocoldevelopment/docs/recist_guideline.pdf}{RECIST guidelines}
}
