% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_fu}
\alias{example_fu}
\title{Generate Simulated Survival Data}
\usage{
example_fu(
  enrolres,
  recist,
  seed,
  lambda_censor = 0.3,
  lambda_control = 0.2,
  beta_arm = -0.6,
  beta_prog_status = 0.5,
  ...
)
}
\arguments{
\item{enrolres}{the enrolment result table, from \code{\link[=example_enrol]{example_enrol()}}.}

\item{recist}{the recist table with progression date, from \code{\link[=example_rc]{example_rc()}}.}

\item{seed}{Integer. Random seed for reproducibility (can be \code{NULL}).}

\item{...}{Additional arguments (currently unused).}

\item{censor_min_days}{minimum time in days between last RECIST evaluation and administrative censoring}

\item{censor_max_days}{maximum time in days between last RECIST evaluation and administrative censoring}
}
\value{
A tibble with \code{N} rows and the following columns:
\itemize{
\item \code{subjid}: Subject ID
\item \code{fu_status}: Status of the patient (0 - Alive/Censored , 1 - Dead)
\item \code{fu_date}: Date of latest news (death date if dead, otherwise censoring date)
}
}
\description{
Internal function that simulates survival data, which depend on treatment arm
and date of progression if applicable.
}
\keyword{internal}
