% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_rc}
\alias{example_rc}
\title{Simulate a RECIST dataset}
\usage{
example_rc(
  enrolres,
  num_timepoints = 10,
  p_new_lesions = 0.01,
  p_not_evaluable = 0.01,
  v_bruits_variation_taille_tumeur = 25
)
}
\arguments{
\item{enrolres}{the enrolment result table, from \code{.example_enrol}}

\item{num_timepoints}{Integer. Number of timepoints for each patient.}

\item{p_new_lesions}{Integer. Probability of a new lesion}

\item{p_not_evaluable}{Integer. Probability of a Not Evaluable measurev_bruits_variation_taille_tumeur=25}

\item{v_bruits_variation_taille_tumeur}{Integer. Standard deviation for the evolution of the tumeur size}
}
\value{
A tibble containing the simulated RECIST dataset.
}
\description{
Used in \code{grstat_example()}. \cr
This function generates a synthetic RECIST dataset following the conventions of clinical oncology trials.
It includes patient tumor size measurements over time and categorizes responses according to RECIST criteria.
}
\section{Columns}{

\itemize{
\item \code{subjid}: The patient identifier.
\item \code{arm} and \code{arm3}: The treatment arm for the patient.
\item \code{rctlsum_b}: Baseline tumor size for patients. The size is simulated following a normal distribution with a mean of 50 and a standard deviation of 30. If the result is <10, it is replaced with a value drawn from a uniform distribution between 70 and 180.
\item \code{rctlsum}: The size of the tumor at each time point. The evolution of the tumor is calculated based on the percentage variation in tumor size from the previous time point. This variation is simulated using a uniform distribution between -30 and 30, with added noise (the noise follows a normal distribution with a mean of 0 and a standard deviation of \code{v_bruits_variation_taille_tumeur}).
\item \code{rctlmin}: The minimal tumor size observed so far.
\item \code{rctlresp}: The response associated with the variation in tumor size, following the RECIST criteria.
\item \code{rcntlresp}: The response associated with non-target lesions.
\item \code{rcnew}: The response associated with the appearance of a new lesion.
\item \code{rcresp}: The global response.
\item \code{rcvisit}: The number of visits.
\item \code{rcdt}: The date of the visits.
}
}

\keyword{internal}
