% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{gr_new_project}
\alias{gr_new_project}
\title{Create a clinical research project}
\usage{
gr_new_project(
  path,
  ...,
  trial_name = NULL,
  headers = NULL,
  open = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Destination directory for the project root. Wille be created if needed.}

\item{...}{Dots are ignored. Reserved for future extensions.}

\item{trial_name}{The abbreviated name of the trial.}

\item{headers}{Optional key–value headers to inject at the top of created files}

\item{open}{If \code{TRUE}, opens the new project in RStudio.}

\item{verbose}{If \code{TRUE}, print diagnostics.}
}
\value{
Invisibly, the normalized path to the created project.
}
\description{
Create a clinical research project with a standardized structure:
\preformatted{
├── main.R
├── NEWS.md
├── README.md
└── my_proj.Rproj
├── R
│   ├── init.R
│   ├── read.R
│   ├── check.R
│   ├── description.R
│   ├── population.R
│   ├── graph.R
│   └── report.R
}
}
\section{Structure}{

At the root of the project:
\itemize{
\item \code{README.md}: short project description
\item \code{NEWS.md}: version log
\item \code{main.R}: central script that orchestrates the workflow
\item \verb{<project>.Rproj}: RStudio project
}

In the R folder:
\itemize{
\item \code{init.R}: packages loading and global options
\item \code{read.R}: data import and global variables
\item \code{population.R}: protocol populations
\item Files ranging from \verb{09_xxx} to \verb{15_xxx} hold your analyses
\item \code{check.R}: data checks (e.g., \code{edc_data_warn()})
\item \code{report.R}: report generation (e.g., \code{{{officer}}})
}
}

\examples{
\dontrun{
  headers = c("Statistician"="Dan",
              "Creation date"="2025-01-01")
  gr_new_project(headers=headers, trial_name="MYSTUDY")
}
}
