% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grstat-package.R
\docType{package}
\name{grstat-package}
\alias{grstat}
\alias{grstat-package}
\title{grstat: Clinical Research Tools}
\description{
Gives tools to standardize the output of clinical research description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oncostat.github.io/grstat/}
  \item \url{https://github.com/oncostat/grstat/}
  \item Report bugs at \url{https://github.com/oncostat/grstat/issues/}
}

}
\author{
\strong{Maintainer}: Dan Chaltiel \email{dan.chaltiel@gustaveroussy.fr} (\href{https://orcid.org/0000-0003-3488-779X}{ORCID})

Other contributors:
\itemize{
  \item Team Oncostat (01ed4t417) [copyright holder, funder]
  \item Baptiste Archambaud \email{baptiste.archambaud@gustaveroussy.fr} [contributor]
  \item Charlotte Bargain \email{charlotte.bargain@gustaveroussy.fr} [contributor]
  \item Aldéric Fraslin \email{alderic.fraslin@gustaveroussy.fr} (\href{https://orcid.org/0000-0003-2235-0012}{ORCID}) [contributor]
  \item Nusaïbah Ibrahimi \email{nusaibah.ibrahimi@gustaveroussy.fr} [contributor]
  \item Livia Pierotti \email{livia.pierotti@gustaveroussy.fr} [contributor]
}

}
\keyword{internal}
