% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{grstat_example}
\alias{grstat_example}
\title{Example databases}
\usage{
grstat_example(N = 200, seed = 42, ...)
}
\arguments{
\item{N}{the number of patients}

\item{seed}{the random seed (can be \code{NULL})}

\item{...}{passed on to internal functions. See \code{\link[=example_ae]{example_ae()}} for control over Adverse Events.}
}
\value{
A list of datasets, like in EDCimport.
}
\description{
Example tables, mostly used in examples and tests.
}
