% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORR_table.R
\name{ORR_table}
\alias{ORR_table}
\alias{as_flextable.ORR_table}
\title{Calculate ORR}
\usage{
ORR_table(
  data,
  ...,
  rc_resp = "RCRESP",
  rc_date = "RCDT",
  subjid = "SUBJID",
  confirmed = FALSE,
  show_CBR = FALSE
)

\method{as_flextable}{ORR_table}(x, ..., confirmed = FALSE, show_CBR = FALSE)
}
\arguments{
\item{data}{A dataset containing longitudinal RECIST data in long format.}

\item{...}{unused}

\item{rc_resp}{The column containing the RECIST response (e.g., \code{"CR"}, \code{"PR"}, \code{"SD"}, \code{"PD"}). Default is \code{"RCRESP"}.}

\item{rc_date}{The column containing the assessment date. Default is \code{"RCDT"}.}

\item{subjid}{The column containing the subject ID. Default is \code{"SUBJID"}.}

\item{confirmed}{Logical; if \code{TRUE}, use the cofirmation method to determine the best response. For CR & PR confirmation of response had to be be demonstrated with an assessment 4 weeks or later from the initial response for response. *As stated in the protocol: «For equivocal findings of progression (e.g., very small and uncertain new lesions; cystic changes or necrosis in existing lesions), treatment may continue until the next scheduled assessment». Therefore, some patients had a response after a progressive disease response, in that case best response was examine before and after PD. Scans conducted after initiating new anti-cancer therapy were not included in the ORR analyses.}

\item{show_CBR}{Logical; if \code{TRUE}, show the Clinical Best Response (CBR). CBR was defined as the presence of at least a partial response (PR), complete response (CR), or stable disease (SD) lasting at least six months (using a window of +/-1 month for the RECIST date).}

\item{x}{a dataframe, resulting of \code{ORR_table()}}
}
\value{
a dataframe (\code{ORR_table()}) or a flextable (\code{as_flextable()}).

a formatted flextable
}
\description{
The function \code{ORR_table()} creates a summary table of recist for each possible response.
The resulting dataframe can be piped to \code{as_flextable()} to get a nicely formatted flextable.
}
\examples{
df = grstat_example()
recist = df$recist

ORR_table(recist)

#the resulting data.frame can be customized using the flextable package
library(flextable)
ORR_table(recist) \%>\%
  as_flextable()
#Is also possible to use the confirmation method for the ORR
ORR_table(recist, confirmed = TRUE) \%>\%
  as_flextable()

#Or show the Clinical Benefice Rate
ORR_table(recist, show_CBR = TRUE) \%>\%
  as_flextable()
}
