% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rando_list.R
\name{randomisation_list}
\alias{randomisation_list}
\title{Randomization lists}
\usage{
randomisation_list(n, arms, strata, block.sizes = c(2, 4), ...)
}
\arguments{
\item{n}{total number of patients}

\item{arms}{treatment arms}

\item{strata}{a list of stratification factors (characters)}

\item{block.sizes}{block sizes (should be a multiple of the number of arms)}

\item{...}{past to \link[blockrand:blockrand]{blockrand::blockrand}}
}
\value{
une tibble
}
\description{
Create randomization lists for EDC softwares using \code{blockrand::blockrand()} for multiple strata at once.
Can use random block size.
}
\section{Details}{


Maximum imbalance per strata is calculated as \code{max(block.sizes)/length(arms)}. \cr
Global maximum imbalance is calculated as \code{max_imbalance*n_strata}.
}

\examples{
# randomisation list for 200 patients randomized in 2 treatment
# arms stratified on 3 groups, with blocks of size 4 or 8

strat = list(age=c("<=18m", ">18m"),
             gender=c("Male", "Female"),
             group=c("A", "B", "C"))
rando = randomisation_list(n=200, arms=c("Control", "Treatment"),
                           strata=strat, block.sizes=c(4, 8))
rando
}
