% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recist_check.R
\name{recist_report_html}
\alias{recist_report_html}
\alias{recist_report_xlsx}
\title{Create a RECIST check report (HTML or Excel)}
\usage{
recist_report_html(
  recist_check,
  output_file = "recist_check_{project}_{date_extraction}.html",
  output_dir = "output/check",
  title = "RECIST Check - {project} - {date_extraction}",
  open = FALSE
)

recist_report_xlsx(
  recist_check,
  output_file = "recist_check_{project}_{date_extraction}.xlsx",
  output_dir = "output/check",
  open = FALSE
)
}
\arguments{
\item{recist_check}{the result of \code{\link[=check_recist]{check_recist()}}}

\item{output_file}{the report file name.}

\item{output_dir}{the directory of output.}

\item{title}{the HTML report title.}

\item{open}{whether to open the report afterward.}
}
\value{
\code{output_file} invisibly. Called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Generate a report as an HTML or Excel file base on RECIST checks made by \code{\link[=check_recist]{check_recist()}}.
}
\examples{
# we unfortunately cannot provide a flawed simulated recist dataset, at least not yet
\dontrun{
db = read_database()
mapping = gr_recist_mapping()
supp_cols_df = enrolres \%>\% select(SUBJID, SITEC, STNO)
recist_check = check_recist(db$rc, mapping=mapping, supp_cols_df=supp_cols_df)
recist_check
recist_report_html(recist_check)
recist_report_xlsx(recist_check)
}

}
