% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{Generate a waterfall plot}
\usage{
waterfall_plot(
  data_recist,
  rc_sum = "RCTLSUM",
  rc_resp = "RCRESP",
  rc_date = "RCDT",
  type = c("best_resp", "worst_resp"),
  rc_star = NULL,
  arm = NULL,
  subjid = "SUBJID",
  warnings = getOption("grstat_wp_warnings", TRUE)
)
}
\arguments{
\item{data_recist}{recist dataset}

\item{rc_sum}{name of the target lesions length sum column in \code{data_recist}, usually "RCTLSUM".}

\item{rc_resp}{name of the response column in \code{data_recist}, usually "RCRESP".}

\item{rc_date}{name of the date column in \code{data_recist}, usually "RCDT".}

\item{type}{one of \code{c("best_resp", "worst_resp")}}

\item{rc_star}{name of the column in \code{data_recist} that triggers the star symbol. The column should be character or factor, with \code{NA} for patients without symbol.}

\item{arm}{name of the treatment column in \code{data_recist}. Can be left to \code{NULL} to not group.}

\item{subjid}{name of the subject identifier column in \code{data_recist}.}

\item{warnings}{whether to warn about any problems}
}
\value{
a ggplot
}
\description{
Generate a waterfall plot
}
\section{Methods}{

Data are ordered on \code{rc_date}.
}

\examples{

\dontrun{
waterfall_plot(rc, rc_date="RCDT", rc_sum="RCTLSUM", rc_resp="RCRESP")
waterfall_plot(rc, rc_date="RCDT", rc_sum="RCTLSUM", rc_resp="RCRESP", type="worst_resp")
rc \%>\%
  left_join(enrolres, by="SUBJID") \%>\% #adds the ARM column
  mutate(new_lesion = ifelse(RCNEW=="1-Yes", "New lesion", NA)) \%>\%
  waterfall_plot(rc_date="RCDT", rc_sum="RCTLSUM", rc_resp="RCRESP",
                 arm="ARM", rc_star="new_lesion")
}
}
